Move Agent process management related classes into their own package
Was in the `services` package which didn't really make sense anymore. This provides better separation and enables other classes to be added without them polluting the services package.

Necessitated a new auto configuration class to encapsulate the module.
