+++ genie-agent/src/main/java/com/netflix/genie/agent/execution/process/JobProcessManager.java
@@ -15,9 +15,10 @@
  *     limitations under the License.
  *
  */
-package com.netflix.genie.agent.execution.services;
+package com.netflix.genie.agent.execution.process;
 
 import com.netflix.genie.agent.execution.exceptions.JobLaunchException;
+import com.netflix.genie.agent.execution.services.KillService;
 import com.netflix.genie.common.dto.JobStatus;
 import org.springframework.context.ApplicationListener;
