+++ genie-agent/src/main/java/com/netflix/genie/agent/execution/statemachine/actions/StateMachineActionsAutoConfiguration.java
@@ -21,11 +21,11 @@ import com.netflix.genie.agent.AgentMetadata;
 import com.netflix.genie.agent.cli.ArgumentDelegates;
 import com.netflix.genie.agent.cli.JobRequestConverter;
 import com.netflix.genie.agent.execution.ExecutionContext;
+import com.netflix.genie.agent.execution.process.JobProcessManager;
 import com.netflix.genie.agent.execution.services.AgentFileStreamService;
 import com.netflix.genie.agent.execution.services.AgentHeartBeatService;
 import com.netflix.genie.agent.execution.services.AgentJobKillService;
 import com.netflix.genie.agent.execution.services.AgentJobService;
-import com.netflix.genie.agent.execution.services.JobProcessManager;
 import com.netflix.genie.agent.execution.services.JobSetupService;
 import com.netflix.genie.common.internal.services.JobArchiveService;
 import org.springframework.context.annotation.Bean;
