+++ genie-agent/src/test/groovy/com/netflix/genie/agent/execution/statemachine/actions/MonitorJobActionSpec.groovy
@@ -21,8 +21,8 @@ import com.netflix.genie.agent.cli.ArgumentDelegates
 import com.netflix.genie.agent.execution.CleanupStrategy
 import com.netflix.genie.agent.execution.ExecutionContext
 import com.netflix.genie.agent.execution.exceptions.ChangeJobStatusException
+import com.netflix.genie.agent.execution.process.JobProcessManager
 import com.netflix.genie.agent.execution.services.AgentJobService
-import com.netflix.genie.agent.execution.services.JobProcessManager
 import com.netflix.genie.agent.execution.services.JobSetupService
 import com.netflix.genie.agent.execution.statemachine.Events
 import com.netflix.genie.common.dto.JobStatus
