+++ genie-agent/src/test/java/com/netflix/genie/agent/configs/ProcessAutoConfigurationTest.java
@@ -1,0 +1,51 @@
+/*
+ *
+ *  Copyright 2019 Netflix, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ *
+ */
+package com.netflix.genie.agent.configs;
+
+import com.netflix.genie.agent.execution.process.JobProcessManager;
+import com.netflix.genie.agent.execution.process.impl.JobProcessManagerImpl;
+import org.assertj.core.api.Assertions;
+import org.junit.Test;
+import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.test.context.runner.ApplicationContextRunner;
+
+/**
+ * Tests for {@link ProcessAutoConfiguration}.
+ *
+ * @author tgianos
+ * @since 4.0.0
+ */
+public class ProcessAutoConfigurationTest {
+    private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
+        .withConfiguration(
+            AutoConfigurations.of(ProcessAutoConfiguration.class)
+        );
+
+    /**
+     * Make sure when the auto configuration is applied it creates the expected beans.
+     */
+    @Test
+    public void expectedBeansExist() {
+        this.contextRunner.run(
+            context -> {
+                Assertions.assertThat(context).hasSingleBean(JobProcessManagerImpl.class);
+                Assertions.assertThat(context).hasSingleBean(JobProcessManager.class);
+            }
+        );
+    }
+}
